+++ src/Resources/assets/js/contao-menu-bundle-init.js
@@ -1,0 +1,5 @@
+require('../scss/contao-menu-bundle.scss');
+
+import { MenuBundle } from './contao-menu-bundle';
+
+document.addEventListener('DOMContentLoaded', MenuBundle.init);