+++ src/Resources/contao/languages/de/tl_module.php
@@ -9,7 +9,5 @@ $lang['huhMenu_maxTriggerLevel'][0] = 'Stoplevel für das Einklappen';
 $lang['huhMenu_maxTriggerLevel'][1] = 'Bis zu dieser Hierarchie-Ebene werden Untermenüs eingeklappt.';
 $lang['huhMenu_openDelay'][0]       = 'Verzögerung beim Aufklappen';
 $lang['huhMenu_openDelay'][1]       = 'Geben Sie hier bei Bedarf eine hier eine Zeit in ms an.';
-$lang['huhMenu_switchDelay'][0]     = 'Verzögerung beim Umschalten zwischen 2 Untermenüs';
-$lang['huhMenu_switchDelay'][1]     = 'Geben Sie hier bei Bedarf eine hier eine Zeit in ms an.';
 $lang['huhMenu_closeDelay'][0]      = 'Verzögerung beim Zuklappen';
 $lang['huhMenu_closeDelay'][1]      = 'Geben Sie hier bei Bedarf eine hier eine Zeit in ms an.';