+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.0.3] - 2020-07-06
-- fixed asset loading for contao 4.9
+## [0.1.0] - 2020-09-18
+- added events
+- added an "open" css class to the menu wrapper if menu opened
+- removed unused fields
+- removed npm package